Supply writable fields only during tag creation
If the whole `Tag` structure is serialized it ends up defining all
(writable) fields. Instead the server should apply defaults where the
client doesn't want them, e.g. the color. To make this possible a new
data structure is introduced per object type in which only specific
fields are defined. The same could be made to work by using pointers in
a structure, albeit that'd be harder to use.